# REF: Make concat not stateful.