--- a/pandas/core/reshape/concat.py
+++ b/pandas/core/reshape/concat.py
@@ -18,7 +18,10 @@ import numpy as np
 from pandas._libs import lib
 from pandas.util._exceptions import find_stack_level
 
-from pandas.core.dtypes.common import is_bool
+from pandas.core.dtypes.common import (
+    is_bool,
+    is_scalar,
+)
 from pandas.core.dtypes.concat import concat_compat
 from pandas.core.dtypes.generic import (
     ABCDataFrame,
@@ -726,14 +729,12 @@ def _clean_keys_and_objs(
         if keys is None:
             keys = objs.keys()
         objs = [objs[k] for k in keys]
-    elif isinstance(objs, (ABCSeries, ABCDataFrame, str)):
+    elif isinstance(objs, (ABCSeries, ABCDataFrame)) or is_scalar(objs):
         raise TypeError(
             "first argument must be an iterable of pandas "
             f'objects, you passed an object of type "{type(objs).__name__}"'
         )
-    elif isinstance(objs, (list, tuple, np.ndarray)):
-        pass
-    else:
+    elif not isinstance(objs, abc.Sized):
         objs = list(objs)
 
     if len(objs) == 0: