… pandas/tests/io/data/excel/test_boolean_types.xlsx