All the logic used to be held in a `_Concatenator` class. The statefullness wasn't really needed since `concat` is a one-shot method so broke it out into its functions