@mroeschke 2024-07-05
```bash
| Change   | Before [039edee0] <main>   | After [0becd77e] <ref/concat>   |   Ratio | Benchmark (Parameter)                                                                |
|----------|----------------------------|---------------------------------|---------|--------------------------------------------------------------------------------------|
| -        | 46.8±4μs                   | 41.8±0.1μs                      |    0.89 | join_merge.ConcatIndexDtype.time_concat_series('string[pyarrow]', 'has_na', 0, True) |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```