@martinfleis 2024-07-06
Hey, just a note that this has [broken](https://github.com/geopandas/geopandas/actions/runs/9817406953/job/27108741741) concat override in geopandas, where we're now hitting `'object' object has no attribute 'objs'`. Just wanted to clarify if that is expected or just an unwanted consequence of refactoring (i.e. whether I should fix it on geopandas side).