@mroeschke 2024-07-06
It was an intended consequence of the refactor given that `concat` isn't stateful anymore. I think I can come up with a fix that will make this non-breaking for geopandas

To confirm it appears that geopandas overrides `__finalize__`? Just curious the reason for that