@martinfleis 2024-07-06
> To confirm it appears that geopandas overrides __finalize__? Just curious the reason for that

To ensure that we can correctly propagate custom metadata. That used to be an information about projection and an active geometry column, right now it is only active geometry column. Without that, the resulting GeoDataFrame would lose the track of active geometry and require manual re-assignment of active geometry.